==> old_experiments_take2/01_a_first_look_at_librispeech.ipynb <==
Let's begin by downloading the librispeech dataset.

!mkdir -p data

fnames = [
    'dev-clean.tar.gz',
    'dev-other.tar.gz',
    'test-clean.tar.gz',
    'test-other.tar.gz',
    'train-clean-100.tar.gz',
    'train-clean-360.tar.gz',
    'train-other-500.tar.gz',    
]
    
for fn in fnames:
    !wget http://www.openslr.org/resources/12/{fn} -P data

%%capture

for fn in fnames:
    !tar -xvf data/{fn} -C data

Audio recordings were aligned with text using the [Montreal Forced Aligner](https://montreal-forced-aligner.readthedocs.io/). We will train our embeddings in an unsupervised way (meaning the text labels corresponding to audio will not be used) however we still want to have the ability to chunk the audio into meaningful parts (we want a complete utterance to constitute an exampole).

Let's obtain labels from this very helpful [repository](https://github.com/CorentinJ/librispeech-alignments). Unfortunately since they are stored on google drive you will need to download them manually on your workstation and upload to the `data` directory in the root of this repository.

%%capture

!unzip data/LibriSpeech-Alignments.zip -d data

import numpy as np
import librosa
import os
from IPython.lib.display import Audio

To read the data, we can adapt the parser from this repository from which we obtained the labels.

librispeech_root = "data/LibriSpeech"

def split_audio(audio_fpath, words, end_times):
    # Load the audio waveform
    sample_rate = 16000     # Sampling rate of LibriSpeech 
    wav, _ = librosa.load(audio_fpath, sample_rate)
    
    start_times = np.array([0.0] + end_times[:-1])
    end_times = np.array(end_times)
    assert len(words) == len(end_times) == len(start_times)
    assert words[0] == '' and words[-1] == ''
        
    segments = []
    segment = []
    for word, st, et in zip(words, start_times, end_times):
        if word == '':
            if not segment: continue
            else:
                segments.append(segment)
                segment = []
        else:
            utterance = wav[int(st*sample_rate):int(et*sample_rate)]
            segment.append([word, utterance])
    return segments

Let's generate train examples and store them on disk. This way our code for training the model will be more streamlined than if we wanted to do something fancy with indexing in a wave file. Simpler code - smaller chance of introducing a bug!

Plus this gives us a chance to offload some computation to before training, which is always nice.

To extract MFCC features, we will use functionality from the [python_speech_features](https://github.com/jameslyons/python_speech_features) repository.

from python_speech_features.base import mfcc
import pandas as pd
import librosa
import uuid

!mkdir -p data/examples
!mkdir -p data/examples_audio

def store_audio(audio):
    fn = uuid.uuid4().hex

    pd.to_pickle(mfcc(audio), f'data/examples/{fn}.pkl')
    librosa.output.write_wav(f'data/examples_audio/{fn}.wav', audio, 16000)
    return fn

%%time

source_words, target_words, source_fns, target_fns, set_names, speaker_ids, book_ids, distances_from_target, audio_fpaths = \
[], [], [], [], [], [], [], [], []

window_size = 3
offsets = list(range(-window_size+1, window_size))
offsets.remove(0)

for set_name in os.listdir(librispeech_root):
    if set_name not in ['train-clean-360', 'train-clean-100', 'test-clean', 'dev-clean']: continue
    set_dir = os.path.join(librispeech_root, set_name)
    if not os.path.isdir(set_dir):
        continue
    for speaker_id in os.listdir(set_dir):
        speaker_dir = os.path.join(set_dir, speaker_id)
        for book_id in os.listdir(speaker_dir):
            book_dir = os.path.join(speaker_dir, book_id)
            alignment_fpath = os.path.join(book_dir, "%s-%s.alignment.txt" % 
                                           (speaker_id, book_id))
            
            if not os.path.exists(alignment_fpath):
                raise Exception("Alignment file not found. Did you download and merge the txt "
                                "alignments with your LibriSpeech dataset?")

            with open(alignment_fpath, "r") as alignment_file:
                for line in alignment_file:
                    # Retrieve the utterance id, the words as a list and the end_times as a list
                    utterance_id, words, end_times = line.strip().split(' ')
                    words = words.replace('\"', '').split(',')
                    end_times = [float(e) for e in end_times.replace('\"', '').split(',')]
                    audio_fpath = os.path.join(book_dir, utterance_id + '.flac')
                    
                    segments = split_audio(audio_fpath, words, end_times)
                    for segment in segments:
                        segment_processed = [[word, store_audio(audio)] for word, audio in segment]

                        for i, (word, path) in enumerate(segment_processed):
                            for offset in offsets:
                                if i + offset < 0 or i + offset > len(segment_processed)-1: continue

                                source_words.append(word), target_words.append(segment_processed[i+offset][0]),
                                source_fns.append(path), target_fns.append(segment_processed[i+offset][1])
                                set_names.append(set_name), speaker_ids.append(speaker_id), book_ids.append(book_id),
                                distances_from_target.append(abs(offset)), audio_fpaths.append(audio_fpath)

len(target_words)

np.unique(set_names)

df = pd.DataFrame(data={
    'source_word': source_words,
    'target_word': target_words,
    'source_fn': source_fns,
    'target_fn': target_fns,
    'set_name': set_names,
    'speaker_id': speaker_ids,
    'book_id': book_ids,
    'distance_from_target': distances_from_target,
    'audio_fpath': audio_fpaths
})

df.shape

df.head()

df.to_csv('data/examples.csv', index=False)
