==> old_experiments_take2/04_analyze_vocab_of_speech2vec_trained_embeddings.ipynb <==
One of the mysteries of how speech2vec was trained is what vocab did the authors use?

There is no information on that provided in the paper, but we can work backwards from the trained embeddings that the authors share!

This will be the highlight of this notebook.

from fastai.vision.all import *
from fastai.learner import *
from fastai.data.all import *
from fastai.callback.tracker import SaveModelCallback
import pandas as pd
import matplotlib.pyplot as plt
from pathlib2 import Path
import numpy as np
import random
from torch.nn import MSELoss

%%time

df = pd.read_csv('data/examples_with_length.csv')
df.shape

import logging
from six import iteritems
from web.datasets.similarity import fetch_MEN, fetch_WS353, fetch_SimLex999
from web.embeddings import fetch_GloVe
from web.evaluate import evaluate_similarity
from web.embedding import Embedding, Vocabulary
from gensim.models import Word2Vec
from gensim.models import KeyedVectors

speech2vec = KeyedVectors.load_word2vec_format('../speech2vec-pretrained-vectors/speech2vec/50.vec', binary=False) 
speech2vec_embeddings = Embedding(Vocabulary(list(speech2vec.vocab.keys())), speech2vec.vectors)

sv_vocab = list(speech2vec.vocab.keys())

len(sv_vocab)

vocab = list(df.source_word.str.lower().unique())

len(vocab)

set(vocab) - set(sv_vocab)

df['in_speech2vec_vocab'] = (df.source_word.str.lower().isin(set(sv_vocab)) & df.target_word.str.lower().isin(set(sv_vocab)))

df.in_speech2vec_vocab.mean()

df.to_csv('data/examples_with_length_speech2vec_vocab.csv', index=False)
